python/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
cssselect==1.1.0
fake-useragent==0.1.11
feedparser==6.0.8
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
parse==1.19.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
yahoo-fin==0.8.9.1
zipp==3.6.0
